avoid circ dependencies by separating states from views
