+++ web/src/state/Window/index.ts
@@ -1,2 +1,2 @@
 export * from "../../views/WithSelectedDiceKey/SelectedDiceKeyViewState"
-export * from "../../views/WindowTopLevelNavigationState"
+export * from "../WindowTopLevelNavigationState"
