+++ web/src/views/LoadingDiceKeys/LoadDiceKeyView.tsx
@@ -3,40 +3,14 @@ import {
   ScanDiceKeyView
 } from "./ScanDiceKeyView";
 import {
-  EnterDiceKeyView, EnterDiceKeyState
-} from "./EnterDiceKeyView"
-import { DiceKeyWithKeyId } from "../../dicekeys/DiceKey";
-import { action, makeAutoObservable } from "mobx";
+  EnterDiceKeyView} from "./EnterDiceKeyView"
 import { observer } from "mobx-react";
 import { CenteredControls, CenterColumn, Instruction, Spacer } from "../basics";
 import { PushButton } from "../../css/Button";
 import { SimpleTopNavBar } from "../../views/Navigation/SimpleTopNavBar";
 import { WindowRegionBelowTopNavigationBarWithSideMargins } from "../Navigation/NavigationLayout";
-import { ViewState } from "../../state/core/ViewState";
-import { NavigationPathState } from "../../state/core/NavigationPathState";
-import { PathStrings } from "../../views/Navigation/PathStrings";
-
-type Mode = "camera" | "manual";
-
-export const LoadDiceKeyViewStateName = PathStrings["LoadDiceKey"];
-export type LoadDiceKeyViewStateName = typeof LoadDiceKeyViewStateName;
-export class LoadDiceKeyViewState implements ViewState {
-  readonly viewName = LoadDiceKeyViewStateName;
-
-  mode: Mode;
-  enterDiceKeyState = new EnterDiceKeyState()
-
-  setMode = action( (mode: Mode) => {
-    this.mode = mode;
-  });
-
-  navState: NavigationPathState;
-  constructor(parentNavState: NavigationPathState = NavigationPathState.root, mode: Mode = "camera") {
-    this.mode = mode;
-    this.navState = new NavigationPathState(parentNavState, LoadDiceKeyViewStateName)
-    makeAutoObservable(this);
-  }
-}
+import { LoadDiceKeyViewState, Mode } from "./LoadDiceKeyViewState";
+import { DiceKeyWithKeyId } from "../../dicekeys/DiceKey";
 
 type LoadDiceKeyProps = {
   onDiceKeyReadOrCancelled: (diceKey: DiceKeyWithKeyId | undefined, howRead: Mode | "cancelled") => any,
