+++ web/src/views/SaveAndDeleteDiceKeyView.tsx
@@ -2,16 +2,14 @@ import { observer } from "mobx-react";
 import React from "react";
 import { CenteredControls, ThisDiceKey, Instruction, Spacer } from "./basics";
 import { OptionButton } from "../css/Button";
-import { DiceKeyWithKeyId } from "../dicekeys/DiceKey";
 import { DiceKeyMemoryStore } from "../state";
 import { NavigationBarForDiceKey } from "./WithSelectedDiceKey/SelectedDiceKeyNavigationBar";
 import { HeightBelowTopNavigationBar, WindowRegionBelowTopNavigationBar } from "./Navigation/NavigationLayout";
 import styled from "styled-components";
 import { cssCalcTyped } from "../utilities";
 import { DiceKeyView } from "./SVG/DiceKeyView";
-import { ViewState } from "../state/core/ViewState";
-import { NavigationPathState } from "../state/core/NavigationPathState";
 import { addressBarState } from "../state/core/AddressBarState";
+import { DeleteDiceKeyViewState, SaveDiceKeyViewState } from "./SaveOrDeleteDiceKeyViewState";
 
 export const ContentRegion = styled(WindowRegionBelowTopNavigationBar)`
   flex: 0 0 auto;
@@ -23,26 +21,6 @@ export const ContentRegion = styled(WindowRegionBelowTopNavigationBar)`
 const IdealMinimumContentMargin = `2rem`;
 const diceKeySize = `min(100vw - 2 * ${IdealMinimumContentMargin}, ( ${HeightBelowTopNavigationBar} - (${IdealMinimumContentMargin}) )/3)` as const;
 const calcDiceKeySize = cssCalcTyped(diceKeySize);
-
-export const SaveDiceKeyViewStateName = "save";
-export const DeleteDiceKeyViewStateName = "delete";
-export type SaveDiceKeyViewStateName = typeof SaveDiceKeyViewStateName;
-export type DeleteDiceKeyViewStateName = typeof DeleteDiceKeyViewStateName;
-export type SaveOrDeleteDiceKeyStateName = SaveDiceKeyViewStateName | DeleteDiceKeyViewStateName;
-export class SaveOrDeleteDiceKeyViewState<SAVE_OR_DELETE extends SaveOrDeleteDiceKeyStateName=SaveOrDeleteDiceKeyStateName> implements ViewState {
-  readonly navState: NavigationPathState;
-
-  constructor(
-    public readonly viewName: SAVE_OR_DELETE,
-    parentNavState: NavigationPathState,
-    public readonly diceKey: DiceKeyWithKeyId,
-    localPath: string | (() => string) = `${viewName}/${diceKey.centerLetterAndDigit}`
-  ) {
-    this.navState = new NavigationPathState(parentNavState, localPath);
-  }
-}
-export type SaveDiceKeyViewState = SaveOrDeleteDiceKeyViewState<SaveDiceKeyViewStateName>;
-export type DeleteDiceKeyViewState = SaveOrDeleteDiceKeyViewState<DeleteDiceKeyViewStateName>;
 
 export const SaveDiceKeyToDeviceStorageContentView = observer( ( {
   state,
